Implement `--git-reference`, remove `--reference`
The `--reference` option is undocumented. I suspect it was accidentally
renamed during the big configure refactor. This commit brings it back
with its original name, but changes behaviour a bit. Now the reference
repository is not only used, but also cloned and fetched. As a result the
user of this option does not need to know how the reference repositories
have to be named or what their URLs are.