Workaround for nmake detection failure
The nmake detection fails randomly on Windows. It does so much more often
on CI systems. The detection fails, because the STDOUT of the make call is
partly or completely lost. This points to a bug in IPC::Cmd, IPC::Open3 or
the Perl core. Instead of trying to hunt down a tricky and long standing
bug in Perl (bug reports of possibly related issues date back to 2014:
https://rt.cpan.org/Public/Bug/Display.html?id=91784) we work around the
issue by trying the detection several times before giving up.

This fixes #19.